Refactor: logic to describe the forwarding path
